pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "kafkaadmin"
version = "0.1.0"
description = "Admin instruments for kafka"
license = "MIT"
authors = ["Vladimir Kamarzin <vvk@vvk.pp.ru>"]
readme = "README.md"
repository = "https://github.com/bitphage/kafka-admin-tool"

keywords = []

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.6.1"
kafka-python = "^2.0.2"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.2.0"
pytest = "^5.4"
pytest-cov = "^2.7"
pytest-mock = "^3.1.0"

sphinx = "^2.2"
sphinx-autoapi = "^1.3.0"
# used in docs to include markdown files
m2r = "^0.2"
# used in docs to parse pyproject.toml
tomlkit = "^0.6"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \venv
  | build
  | dist
)/
'''
